autoconf: don't remove disconnected hosts from nagios
 Currently auto-config removes all the hosts with disconnected
status in 'peer status' from nagios configuration. But that
is not correct as a host can goto disconnected status for
many differnt reasons and we should not remove the configuration.

 Similarly disconnected hosts need not be added to nagios
configuration during auto discovery.

  Fixing this by ingnoring all the disconnected hosts in delta
calculation.

Note: Host name can't be used to identify disconnected hosts
as we may not be able to communicate the disconnected host
to get the host name. Hence Host UUID in peer status is
used to intentify the host.

Bug-Url: https://bugzilla.redhat.com/1091170

Change-Id: I3fe730d2545f8cbc8224c576712b09bb4c16e712
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/7677
Reviewed-by: Sahina Bose <sabose@redhat.com>