nagios-server-addons: NRPE command with timeout
Introduced a utility method to return a NRPE base command
with timeout set externally.

Currently if a plugin internally executes NRPE to get the
details from the node, there is no mechanism that timeout
could be set to more than 10 sec. This method provides the
NRPE command with timeout (if passed).

This is required for some of the NRPE calls where gluster commands
get executed and they might take more time.

All the plugins, which execute a NRPE within, can provide
optional command line argument for timeout, and same can be used
for forming the proper NRPE call with timeout value.

Change-Id: Id97624df743664a320a585acc4a85cfcf64d0a07
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>
Reviewed-on: http://review.gluster.org/7682
Reviewed-by: Sahina Bose <sabose@redhat.com>
Tested-by: Sahina Bose <sabose@redhat.com>