autoconf: validate the nagios server address in auto-config
  Currently Nagios server address entered by the user during
auto config was not being verified. This patch helps to verify
the address entered by the user.

  If IP address is given as the nagios server address then it
checks the pattern and verifies that it is mapped to one of the
non loopback device in the host

  If user enters fqdn name, then it tries to resolve it,
also it verifies that resolved IP address maps to one of
the non loopback device in the host.

Bug-Url: https://bugzilla.redhat.com/1127657
Change-Id: I88d67cc6d8fa05f2934922fbc0d8e757b1d73e43
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/7740
Reviewed-by: darshan n <dnarayan@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>