nagios-server-addons: Fix for dummy config file issue.
This patch replaces the configfile node1.cfg(gluster-host)
by temp_node1.cfg which is of type linux-server having
service PING. And this dummy config file is renamed to
temp_node1.cfg.sample during auto-discovery as it is not
needed after we configure actual hosts. Also this patch
sets the config variable allow_empty_hostgroup_assignment
to 1, which is needed for nagios to allow declaration of
empty host-group with services.

Change-Id: I03cc54975f5ff5857f8d82b66110e049e4195009
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1091236
Signed-off-by: ndarshan <dnarayan@redhat.com>
Reviewed-on: http://review.gluster.org/7952
Reviewed-by: Kanagaraj M <kmayilsa@redhat.com>
Reviewed-by: Shubhendu Tripathi <shtripat@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>