autoconfig: avoid renaming hostnames in nagios
  Auto Config tries to use the output of 'hostname' command as
a host_name for node in nagios. But if the hostnames are not unique,
then it uses IP Address as node_name name in nagios.

 If after removal of a node,  hostnames becomes unique then
auto config renames all the host configuration in nagios to their
hostnames instead of ip address. It creates multiples issues like
removing all the nodes, loss of time series data and unnecessary
confusion to the user.

 Fixing this issue by retaining the perviously configured host
names always by going through the host uuid. So if duplication
comes after a node addition then only that node will be
configured with ip address as host_name but all other existing
hosts will not be touched.

 Similary, even though there is no duplication after the removal
of a node with duplicate name, auto config will retain the
previously configured hostnames (which are ip addresss
becasue there was a hostname duplication earlier) instead of
renaming.

Change-Id: I0abef47e2e495a591dd3901d2b1766a04c59a701
Bug-Url: https://bugzilla.redhat.com/1139228
Bug-Url: https://bugzilla.redhat.com/1139505
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/8699
Reviewed-by: darshan n <dnarayan@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>