server-plugin: consider all status of volume in cluster status
Consider the 'UNKNOWN' status of volumes while calculating the cluster
status. Following will be the cluster state and state information.

Cluster State                State Information
UP      "OK : None of the Volumes in the cluster are in Critical State"
UP      "OK : No Volumes present in the cluster"
UP      "WARNING : Some Volumes in the cluster are in Critical State"
UP      "WARNING : Some Volumes in the cluster are in Unknown State"
UP      "WARNING : Some Volumes in the cluster are in Warning State"
UP      "WARNING : All Volumes in the cluster are in Warning State"
DOWN    "CRITICAL: All Volumes in the cluster are in Critical State"
DOWN    "CRITICAL: All Volumes in the cluster are in Unknown State"

Bug-url: https://bugzilla.redhat.com/1128007
Change-Id: I06fb1697cb4919420ab6d6ea54e9d9ee96820d2a
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/9053
Reviewed-by: Shubhendu Tripathi <shtripat@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>